Бонусные абонементы по промокодам

Closes #41

See merge request !114

+++ src/main/java/ru/majordomo/hms/personmgr/controller/rest/AccountAbonementRestController.java
@@ -3,7 +3,6 @@ package ru.majordomo.hms.personmgr.controller.rest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
-import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
@@ -13,22 +12,19 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
-import java.time.LocalDateTime;
-import java.time.Period;
-import java.util.HashMap;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Map;
-import java.util.Optional;
 
+import ru.majordomo.hms.personmgr.common.AbonementType;
 import ru.majordomo.hms.personmgr.exception.ParameterValidationException;
 import ru.majordomo.hms.personmgr.model.PersonalAccount;
 import ru.majordomo.hms.personmgr.model.abonement.Abonement;
 import ru.majordomo.hms.personmgr.model.abonement.AccountAbonement;
+import ru.majordomo.hms.personmgr.repository.AbonementRepository;
 import ru.majordomo.hms.personmgr.repository.AccountAbonementRepository;
 import ru.majordomo.hms.personmgr.repository.PersonalAccountRepository;
-import ru.majordomo.hms.personmgr.repository.PlanRepository;
 import ru.majordomo.hms.personmgr.service.AbonementService;
-import ru.majordomo.hms.personmgr.service.FinFeignClient;
 import ru.majordomo.hms.personmgr.validators.ObjectId;
 
 @RestController
@@ -38,16 +34,19 @@ public class AccountAbonementRestController extends CommonRestController {
 
     private final PersonalAccountRepository accountRepository;
     private final AccountAbonementRepository accountAbonementRepository;
+    private final AbonementRepository abonementRepository;
     private final AbonementService abonementService;
 
     @Autowired
     public AccountAbonementRestController(
             PersonalAccountRepository accountRepository,
             AccountAbonementRepository accountAbonementRepository,
-            AbonementService abonementService) {
+            AbonementService abonementService,
+            AbonementRepository abonementRepository) {
         this.accountRepository = accountRepository;
         this.accountAbonementRepository = accountAbonementRepository;
         this.abonementService = abonementService;
+        this.abonementRepository = abonementRepository;
     }
 
     @RequestMapping(value = "/{accountAbonementId}", method = RequestMethod.GET)
@@ -72,13 +71,14 @@ public class AccountAbonementRestController extends CommonRestController {
 
         AccountAbonement accountAbonement = accountAbonementRepository.findByIdAndPersonalAccountId(accountAbonementId, account.getId());
 
-        Boolean autorenew = requestBody.get("autorenew") != null ? Boolean.valueOf(requestBody.get("autorenew")) : false;
-
-        accountAbonement.setAutorenew(autorenew);
-
-        accountAbonementRepository.save(accountAbonement);
-
-        return new ResponseEntity<>(HttpStatus.OK);
+        if (!accountAbonement.getAbonement().isInternal()) {
+            Boolean autorenew = requestBody.get("autorenew") != null ? Boolean.valueOf(requestBody.get("autorenew")) : false;
+            accountAbonement.setAutorenew(autorenew);
+            accountAbonementRepository.save(accountAbonement);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
     @RequestMapping(value = "/{accountAbonementId}", method = RequestMethod.DELETE)
@@ -88,9 +88,13 @@ public class AccountAbonementRestController extends CommonRestController {
     ) {
         PersonalAccount account = accountRepository.findOne(accountId);
 
-        abonementService.deleteAbonement(account, accountAbonementId);
-
-        return new ResponseEntity<>(HttpStatus.OK);
+        AccountAbonement accountAbonement = accountAbonementRepository.findByIdAndPersonalAccountId(accountAbonementId, account.getId());
+        if (!accountAbonement.getAbonement().isInternal()) {
+            abonementService.deleteAbonement(account, accountAbonementId);
+            return new ResponseEntity<>(HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
     }
 
     @RequestMapping(value = "", method = RequestMethod.GET)
@@ -124,7 +128,19 @@ public class AccountAbonementRestController extends CommonRestController {
 
         Boolean autorenew = requestBody.get("autorenew") != null ? Boolean.valueOf(requestBody.get("autorenew")) : false;
 
-        abonementService.addAbonement(account, abonementId, autorenew);
+        Boolean preorder = requestBody.get("preorder") != null ? Boolean.valueOf(requestBody.get("preorder")) : false;
+
+        Abonement abonement = abonementRepository.findOne(abonementId);
+
+        if (abonement == null) {
+            throw new ParameterValidationException("Abonement with abonementId: " + abonementId + " not found");
+        }
+
+        // Internal абонементы юзер не может заказывать
+        if (!(abonement.isInternal()))
+            abonementService.addAbonement(account, abonementId, autorenew, false, preorder);
+        else
+            throw new ParameterValidationException("Interal abonement is not allowed for adding");
 
         return new ResponseEntity<>(HttpStatus.OK);
     }
